Updated screen savers to use new namespace
Necessary to organize these files since they are similar in nature. This is a breaking change that will soon be resolved once a container is added to allow these objects to be injected.

Milestone: patch
